README.md
ConstraintLayout allows you to create large and complex layouts with a flat view hierarchy (no nested view groups). It's similar to
 RelativeLayout in that all views are laid out according to relationships between sibling views and the parent layout, 
but it's more flexible than RelativeLayout and easier to use with Android Studio's Layout Editor.

All the power of ConstraintLayout is available directly from the Layout Editor's visual tools, because the layout API and the Layout
 Editor were specially built for each other. So you can build your layout with ConstraintLayout entirely by drag-and-dropping instead 
of editing the XML.


ConstraintLayout is available in an API library that's compatible with Android 2.3 (API level 9) and higher. This page provides a 
guide to building a layout with ConstraintLayout in Android Studio 3.0 or higher.

https://developer.android.com/training/constraint-layout/